add moderation folder for better code reuse
