+++ server/moderation/banned-usernames.js
